% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinemas.R
\name{shinemas}
\alias{shinemas}
\title{Query SHiNeMaS and format data for specific R packages}
\usage{
shinemas(db = "shinemas.local.seed", id = "user", pwd = "toto",
  query_type = "data_agro", specie = NULL, project = NULL,
  variable = NULL, germplasm = NULL, germplasm_type = NULL,
  relation_type = NULL, location = NULL, year = NULL)
}
\arguments{
\item{db}{url of the database}

\item{id}{id to connect to the database}

\item{pwd}{password to connect to the database}

\item{query_type}{name of the query. Possible values are
\itemize{
 \item PPBstats_data_network_unipart_seed_lot
 \item PPBstats_data_agro
 \item PPBstats_data_SR
 \item PPBstats_data_HA
}}

\item{specie}{filter: specie to keep}

\item{project}{filter: project to keep}

\item{variable}{filter: variable to keep}

\item{germplasm}{filter: germplasm to keep}

\item{germplasm_type}{filter: germplasm type to keep}

\item{relation_type}{filter: relation type to keep}

\item{location}{filter: location to keep}

\item{year}{filter: year to keep}
}
\value{
The function return a dataframe formated for specific query type and packages
}
\description{
\code{shinemas} queries SHiNeMaS and formats data for specific R packages
}
\details{
The database SHiNeMaS (Seed History and Network Management System) is available here: \url{https://sourcesup.renater.fr/frs/?group_id=2295}

The query is based on a webservice ... TO COMPLETE

The query_type is under the format [package-name]_[query-name].
For example PPBstats_data_agro stands for data_agro format for package PPBstats.

PPBstats is an  R package for Participatory Plant Breeding statisticial analyses. 
More information are available here: \url{https://priviere.github.io/PPBstats_web_site/}
}
\examples{
\dontrun{
data_agro = shinemas(
db = "shinemas.local.seed",
id = "user",
pwd = "toto",
query = "PPBstats_data_agro"
)

library(PPBstats)
data = format_data_PPBstats(data_agro, type = "data_agro")
}

}
\author{
Pierre Riviere and Laetitia Courgey
}
